pyproject.toml
[tool.poetry]
name = "connection-svelte-django"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "n"
readme = "README.md"
packages = [{include = "connection_svelte_django"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
whitenoise = "^6.2.0"
Pillow = "^9.2.0"


[tool.poetry.group.dev.dependencies]
black = "^22.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
